Modify Logic, refactor code composition

+++ philo/print.c
@@ -19,7 +19,7 @@ void ft_putstr(char *str)
 
 void safe_write(t_philo *philo, char *msg)
 {
-	unsigned long current_time;
+	long long current_time;
 
 	pthread_mutex_lock(&philo->params->write_mutex);
     current_time = get_elapsed_time((long long)philo->params->start_time);
